requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.3
psycopg2-binary==2.9.9
supabase==2.8.0
redis==5.1.1
requests==2.32.3
httpx<0.28,>=0.24
aiosmtplib==3.0.2
reportlab==4.2.5
qrcode[pil]==7.4.2
pyqrcode==1.2.1
pypng==0.20220715.0
paypalrestsdk==1.13.3
apscheduler==3.10.4
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
Pillow==10.4.0
email-validator==2.2.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
PyJWT==2.9.0
cryptography==43.0.1
